Add accessLocation to DisplayBag
