+++ archive/registrar_http/src/main/scala/uk/ac/wellcome/platform/archive/registrar/http/models/DisplayBag.scala
@@ -1,13 +1,14 @@
 package uk.ac.wellcome.platform.archive.registrar.http.models
 import io.circe.generic.extras.JsonKey
 import uk.ac.wellcome.platform.archive.common.models.DisplayStorageSpace
-import uk.ac.wellcome.platform.archive.registrar.common.models.{BagDigestFile, FileManifest, StorageManifest}
+import uk.ac.wellcome.platform.archive.registrar.common.models._
 
 case class DisplayBag(
   id: String,
   space: DisplayStorageSpace,
   info: DisplayBagInfo,
   manifest: DisplayFileManifest,
+  accessLocation: DisplayLocation,
   createdDate: String,
 
   @JsonKey("type")
@@ -20,32 +21,7 @@ object DisplayBag {
     DisplayStorageSpace(storageManifest.id.space.underlying),
     DisplayBagInfo(storageManifest.id.externalIdentifier.underlying),
     DisplayFileManifest(storageManifest.manifest),
+    DisplayLocation(storageManifest.accessLocation),
     storageManifest.createdDate.toString
   )
-}
-
-case class DisplayBagInfo(
-  externalIdentifier: String,
-  @JsonKey("type")
-  ontologyType: String = "BagInfo"
-                         )
-
-case class DisplayFileManifest(
-  checksumAlgorithm: String,
-  files: List[DisplayFileDigest],
-                                @JsonKey("type")
-ontologyType: String = "FileManifest"
-                              )
-
-object DisplayFileManifest {
-  def apply(fileManifest: FileManifest): DisplayFileManifest = DisplayFileManifest(fileManifest.checksumAlgorithm.value, fileManifest.files.map(DisplayFileDigest.apply))
-}
-
-case class DisplayFileDigest(checksum: String,
-                             path: String,
-  @JsonKey("type")
-                             ontologyType: String = "File")
-
-object DisplayFileDigest {
-  def apply(bagDigestFile: BagDigestFile): DisplayFileDigest = DisplayFileDigest(bagDigestFile.checksum.value, bagDigestFile.path.value)
 }
