+++ archive/registrar_http/src/main/scala/uk/ac/wellcome/platform/archive/registrar/http/models/DisplayBagInfo.scala
@@ -1,0 +1,8 @@
+package uk.ac.wellcome.platform.archive.registrar.http.models
+import io.circe.generic.extras.JsonKey
+
+case class DisplayBagInfo(
+  externalIdentifier: String,
+  @JsonKey("type")
+  ontologyType: String = "BagInfo"
+                         )
